test: move robotjs-based devtools paste test to main runner
